requirements.txt
rich
prompt_toolkit
trilium-py
requests
nest-asyncio
asyncio
gitpython